wip: simplified cleaning functions, working on TF-IDF
