+++ .gitignore
@@ -94,6 +94,7 @@ target/
 
 # pyenv
 .python-version
+.idea/
 
 # celery beat schedule file
 celerybeat-schedule
