+++ occupationcoder/coder/code_matcher.py
@@ -1,0 +1,11 @@
+import os
+import sys
+
+from sklearn.feature_extraction.text import TfidfVectorizer
+from sklearn.metrics.pairwise import cosine_similarity
+
+class MixedMatcher():
+
+    def __init__(self):
+        return None
+
